pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "synthapi"
version = "0.1.0"
description = "A tool for local API mocking"
requires-python = ">=3.7"
dependencies = [
    "typer[all]>=0.9.0",  # Added [all] to include all optional dependencies
    "openai>=1.0.0",      # For GPT-4 API integration
    "python-dotenv>=1.0.0"  # For environment variable management
]

[project.scripts]
synthapi = "synthapi.cli:app"